refactor(check-saved-objects): move name/title validation functions to their respective utils files
- `validateNameTitleFieldTypesNewType` → `new_type_utils.ts` (new file)
- `validateNameTitleFieldTypesExistingType` → `existing_type_utils.ts`
- `getInvalidNameTitleFields` and `isSearchableViaManagement` remain in `common_utils.ts`
  and are now exported so the utils files can import them

Made-with: Cursor
